Filter bus voltage before using it for control
Also, report the bus voltage over the debug channel, up its rate to
545000, and increase a gain now that it is possible.

The bus voltage was being sampled when the low side FETs were on, and
could get relatively large (4-5V) above nominal during that period.
This introduced a feedback loop and destabilized the current
controller a fair amount.