stm32f446_async_uart: Improve performance on bulk transfers
For bulk reads, enable the "half-transfer complete" interrupt, which
will give us an interrupt whenever we're at the midway point of our
circular buffer.  That, combined with the "tcif" flag, will give us
two intermediate points in each buffer to report results to the
client.

For transfers, don't de-assert the direction line if we have another
transfer immediately queued up.