Implement a watchdog for position mode
If enabled, then if valid position commands are not received
frequently enough we enter a "position timeout" mode, which zeros out
the P control term and applies an alternative configurable torque
limit.

The hope is that machines will be more likely to be rendered safe upon
loss of commands if we actively try to keep things slow rather than
just cut power entirely.  Cutting power entirely is still reserved for
other fault types.