By default, capture and restore config across flash
Since the controller doesn't perform any schema mapping when loading
configs, if the schema for any configuration changed, it would be lost
after re-flashing.  We now capture everything as text, then command it
back after the flash so that it all is persisted.

A cmdline flag (--no-restore-config) can disable this behavior.