Increase the ADC prescaler to 4x
Now that I plotted all 3 phase currents during a calibration cycle, it
was obvious the ADCs weren't working correctly.  The sample time was
not the problem after all, but the prescaler was.  Increase the
prescaler, drop the sample time, and now the current waveforms are
nice and smooth across all three phases.