Limit cancelation points in the fdcanusb transport
On windows, the cancelation mechanism used to discard data after
sending "tel stop" annoyingly always seemed to happen right after we
had sent a query to a device, but before having received the reply.
This would cause the fdcanusb transport to get out of sync.  Now we
disallow cancelation in that time window.

Resolves #6