Basic "max_position_slip" functionality
This can be used to implement the equivalent of an anti-windup term on
the position control loop.  When enabled, the control position will be
limited to be no more than a certain amount further away from the
current position.

This means that in pure velocity mode, the controller can be made to
limit the amount of "catch up" that occurs.  When in position mode,
external forces can drag the control set point around and it stays
persistently.