Scale up control_position to save runtime cost
Now servo_stats.control_position is scaled such that it is 2**32 times
bigger than unwrapped_position_raw.  This means that wraparound is
handled in the same way by the hardware, and some operations align to
words better resulting in the overall ISR taking less time.

Previously it would occasionally trip the over-time limit in the
regression test, where that has not been observed with this setup.