Make target autodetection work again
This reverts most of 5f52861dd5eff8ff929ddfd9d925d1cefdda73fd and uses a
different approach, which makes the aiostream class support cancellation
natively, instead of trying to handle it at the fdcanusb level.

Despite that, Windows still did weird things if you tried to schedule a
timeout at 0.01 or 0.02s, so for target detection, we just put it at a
bigger timeout for now.