Implement a maximum velocity limit
When the measured velocity exceeds this value, the allowable output
current is derated, reaching 0 at the top of the derating window.

The default maximum is large, 30000rpm, which is beyond what the
system can support.  The devkit default maximum is 600rpm, so as to
make using the devkit less error prone out of the box.