Make the PWM rate be configurable
* Available options are between 15kHz and 60kHz
* Above 40kHz, the control cycle is run at half the PWM rate

To make this work and still fit within timing constraints, ADC1/2 and
ADC3/4 are configured into dual mode, so that we can start each pair
with one operation.

In order to have PWM compensation work at least passably across the
range of frequencies, we change the default compensation for r4.8
boards and update moteus_tool to force those changes when upgrading
firmware.

The maximum power configurable value is now defined relative to 40kHz,
so that the PWM rate and power can be changed independently.

Update moteus_tool to limit the torque and encoder bandwidth according
to the currently configured pwm rate.