Implement a more comprehensive encoder and aux port system
Each of the two auxiliary ports can now have its pins individually
configured and assigned to many different options for inputs,
including SPI, quadrature, sine/cosine, and hall effects.

The auxiliary port inputs can be configured individually to produce
theta values for commutation, and output position/velocity for
position mode control.  Additionally, flexible options are provided to
use index inputs either for commutation, or as homing switches on the
output.

Position mode control is moved to operate purely in a local relative
coordinate system.  This means that rezeroing and forcing the output
position can now occur without affecting control.  Only the next
command will take into account the new values (aside from optional
absolute position limits).

Position mode control has a new optional per-command override to enter
"fixed voltage" mode, which can be useful to perform homing on
incremental commutation sources.

This revises the ELF version to 0x0105 because of the new
configuration, and the register protocol version to 5, because of the
changed semantics of the "rezero state" register.

moteus_tool attempts to port configuration from old versions to the
new.  Its backporting ability is largely limited to those
configurations which it generated itself as a result of an upgrade.