Improve cooldown behavior
Previously, we would hard brake for some number of cycles during the
cooldown window.  That was a reasonable thing for D current static
stress tests, but not so reasonable if the motor was spinning with
inertia.

Now, we instead remain in current control mode and target 0 D and Q
current while in cooldown.  This gives good results both for
stationary D current tests, and while powering at speed (for which it
mostly does nothing).