Initialize the STM32G4 ADCs in turn, rather than all at once
The datasheet says nothing about simultaneous initialization being
problematic.  However, if the ADEN bits of two ADCs are turned on in
too close proximity to each other, "bad things" happen.  Here, the
observed failure was that the prescaler seemed to be set back to 1,
despite the prescaler register still reading the non-unity value set.

This probably only started being a problem with the switch to gcc-11,
which did a better job of packing these enables into a smaller amount
of time.  Even so, it would only be triggered when the code aligned
properly onto some boundary, probably the flash accelerator fetch
block.  Thus the problem would come and go with the phase of the moon.

The observed failure was that the raw ADC values would have many
spurious readings of exactly 2048 or near 2048.