Support synchronizing the clock of a moteus controller to the host
In the firmware, this exposes two new registers.  The first reports a
counter that increments once per millisecond.  The second is a command
that allows the application to change the processor speed in 0.25%
increments.  Document these new registers and expose them to python.

Implement an example python script that shows how these registers can
be used to force a moteus controller to operate with time synchronized
to the host application.