Ensure that relative position stays in sync with the encoder
The relative position was calculated by measuring the encoder delta,
scaling it by the gear ratio, and then accumulating.  However, due to
floating point error, this conversion is lossy, and if the encoder
goes from reading 1 to 2, and then back to 1, the relative position
will not necessarily end up at the exact same location.

Instead, now, we ensure that the relative position is always equal to
an integer number of encoder revolutions plus the current filtered
encoder value.

This adds roughly 5-8 cycles to the ISR.