Be more robust to startup ordering
In some cases, it was reported that when attempting to enable the
driver immediately after powering on, this assert could trigger.  I
believe with the DRV8353 (but probably not the DRV8323), the
calibration sequence is automatically initiated on startup and can be
reported as being active over SPI.

So for now, just don't even care what it was to begin with, and ensure
that we set all the regs to 0 before leaving.  Add a new fault code
for if that somehow goes wrong.