Improve the semantics of non-zero target velocities
When acceleration or velocity limits were enabled, the previous
semantics were:

 "be at velocity v_f when you reach x_f"

These are challenging semantics to work with, especially if the
initial velocity is non-zero.  For instance, if you were traveling at
+20Hz at x=10, and the new target was v=20Hz, x=9.9999, the resulting
trajectory would have to come to a stop, accelerate backwards, then
ramp up speed in order to reach 20Hz again while passing through
9.9999.

The new semantics are:

 "match the trajectory defined by x=x_f + t * v_f"

This provides a more natural interpretation of the arguments, is
more useful in practice, and requires fewer ISR cycles.