moteus-c1 r1.0
This is a lower-power, lower-cost, smaller version of the moteus-n1.
Rough top-line specs:

* Dimensions: 38x38x9mm
* Voltage: 10-51V
* Continuous phase current: 5A/15A
* Peak phase current: 20A
* Peak power: 100W
* I/O connectors: AUX2
* Board mountable: Power, CAN, AUX2

It uses the same power and CAN connectors as other moteus controllers
(XT30 / JST PH3), and can use the same firmware.  It has the same
power and phase terminal pattern as the moteus-n1 that can fit a
terminal block or a 0.1" pin header.