Make board_debug commands error on malformed values
A common mistake has been to send a command like:

d pos 0 0 f1

Which previously would be erroneously accepted.  The "f1" would be
accepted as a maximum torque, parsed as 0, and no feedforward torque
would be applied at all.  Now we error for all floating point parsing
that is malformed.