Bump the ABI version to 0x0107
This is necessary because old servos with

motor_position.output.sign == -1

had a workaround in moteus_tool that is no longer necessary.  This way
moteus_tool can identify which version of firmware it is dealing with
and do the right thing (if there is really a right thing for the old
version) for both versions.