Interpolate commutation theta offsets
Older versions of moteus_tool can generate offset tables which are not
suitable for interpolation.  Try to catch those and error so the user
knows they need to re-calibrate.

This almost certainly puts our ISR execution time above the dyno test
limits.  Later work will have to optimize the performance of this.