Optionally use non-linear optimization for generating commutation tables
This is only used if the initial fit was not good enough, but enables
operation in systems that have a lot of non-linearity, like off-axis
encoders.  Command line options are available to control whether
optimization is used or not, as it can be relatively slow, especially
on devices like the Raspberry Pi.