Implement PWM output for aux ports
Capable pins can now be configured as a PWM output.  After doing so,
either a "aux[12] pwm" diagnostic mode command or a register write can
be used to set the current PWM value.

Add support in the python and C++ libraries (and add gpio support in
the C++ library while we're at it).