Greatly improve encoder compensation in hard configurations
Switch from 32 bins of compensation to 256.  This is implemented as an
int8 + scale, in the same was as cogging compensation so as to
minimize the amount of memory, flash, and CAN communication required.

Update the encoder_compensation script to support both inertial and
reference mode compensation calculations.  Improve and label all
plots.

Add a `compare_encoder.py` script, which can be used to do ground
truth evaluation with a "reference" high accuracy encoder.

These changes primarily help off-axis configurations, or any encoder
that has a fair amount of predictable non-linearity.