Limit the configured maximum current to the available hw range
The DRV8323/DRV8353 current sense amplifier has a documented linear
range which does not extend rail to rail.  Between that, and the
configured current sense gain, there is a finite region of current
which it is possible to actually sense.  It makes no sense to
configure a maximum current beyond that aside from development
testing, so just hard limit it to that value with a small 5% margin to
leave some room for control.