Improve the power limiting subsystem
Make the power limit be dependent upon both input voltage and PWM
frequency.  Since the power limit is largely determined by DC voltage
ripple, at lower input voltages we both have more headroom, and the DC
bias derating of the MLCC capacitors on moteus controllers gives us
larger effective capacitance.

Change the sense of the `servo.max_power_W`.  It used to configure a
"nominal" power, which was then derated by PWM.  Now, it configures an
absolute power to be applied as an "if lower then" onto the factory
profile for each board.

When limiting power, do so in the current domain rather than the
voltage domain, and do so based on the total current vector, not the D
and Q phases separately.  This is more stable at speed, and does at
least something reasonable when flux braking (or possible future field
weakening) is in effect.