Remove PWM dead time compensation support
After further investigation, the non-linearity is actually related to
zero current and the ensuing recovery of the MOSFETs and their body
diodes, and is not related to the PWM dead time in any significant
way.  Thus while this helped near 0 current, it hindered at non-zero
current and results in incorrect voltages being reported.