Make inductance calibration work reliably for more motors
We previously did an OK job at low resistance / low inductance motors,
and high resistance / high inductance motors, but not so good a job at
high resistance / low inductance ones.

Instead of trying to guess at a single period to check out, we just to
a sweep over periods, attempting to stop if it looks like we've found
the peak so as to limit the peak ripple current when checking very low
inductance motors.  This gives better results across the entire range
of motors.