Do a better job of determining phase resistance
Measure the resistance without curve fitting, just relying on two
extremal points, which are the highest current measured and some
fraction of that.  Do away with the separate stage to first find the
resistance calibration voltage just doing both in the same loop.  This
relies completely on --cal-motor-power to determine how much to apply.

While we are at it, fix the definition of power to actually match
electrical power.  That means changing the default motor cal power to
8.5W (which is what it always was, we just used an incorrect
definition previously).