Make the implementation of --cal-no-update more robust
We would previously carefully attempt to monitor every possible
change, and if changes were always required, attempt to roll them back
if the flag was set.

Instead now, we just capture all configuration when we start, and
stuff them all back in when we are done.  If there is a
mid-calibration error they won't be reset, but they also won't be
written to persistent storage.  This seems like a reasonable tradeoff.