Improve inductance calibration for all motor types
Measure inductance with a voltage square wave that is not centered
around 0V.  There is a large non-linearity around 0 current.  By
offsetting our inductance calibration waveform, we get a much more
accurate result.

To accomplish this, "d ind" now accepts an optional fixed voltage
override and an optional method to ignore bounds (although it is not
required yet).