Improve robustness when operating with PSUs
Re-formulate the flux braking parameters to be in terms of a
difference from the maximum voltage, rather than an absolute value.
In the previous formulation, if the flux braking voltage was ever
lowered, then the system was connected to a higher voltage supply, the
controller would draw arbitrarily large amounts of power due to flux
braking.

Now, the primary knob to turn is the maximum voltage.  If the maximum
voltage is too low, the user will immediately get an over-voltage
fault instead of a uvlo fault 33.

Use the new feature with developer kits, which have a 24V PSU, by
setting the maximum voltage to 30V, which results in a flux braking
threshold of 27V.  With the previous defaults, it was easy to drive
the output of the PSU up to the flux braking threshold of over 50V,
which can easily destroy it or result in catastrophic damage.