Update temperature thresholds to be configured with margin
This, like the flux braking parameter before it, means that changing
the maximum limit temperature is a one stop change rather than two
parameters that need to be updated in tandem.  That should make it
easier to change temperature thresholds correctly.

It requires nearly the same upgrade logic, so we might as well do it
in the same ABI transition.