Report *why* power was limited when limiting has occured
This re-uses the "fault code" register and field in `servo_stats` to
mean something during normal control operation.  The reason being that
everyone was likely already requesting it, and it was previously not
defined when not in the fault mode.

It is possible that some consumers were treating any non-zero fault
code as a fault regardless of the primary mode, in which case this
change will require updates to client side code.