Increase the default timeouts for the C++ transports
The python library has an infinite timeout, so this is still much less
than that!

The shorter timeouts caused problems when reading large diagnostic
streams, as once one thing timed out, we started spamming the
controller with a new request every 5ms, which made it go even slower
and stop responding promptly.