Refactor the interface to the python CAN-FD transports
This updates the interface for the python CAN-FD transports to enable
a few potential improvements down the road.

1. Supporting transparent operation with multiple physical transports
in a high performance "cycle-like" mode.

2. Operate in domains where multiple application streams are running
at the same time, and thus it is not possible to ensure that the
response to any given command will always be the "next" frame to be
received.  This for instance can allow python code to be running at
the same time as the diagnostic channel is being streamed by using the
response filters to ensure that the correct sender gets the correct
responses.

This is intended to be a non-functional change, although it is
possible that integration with the pi3hat is broken here.