Implement UUID based discovery at the transport level
moteus_tool now uses this when no targets are specified to determine
what to communicate with, and report a more useful error when there
are conflicts.  (It still only addresses devices by CAN-ID for now)

This required factoring various pieces of functionality out of
moteus.py so that they can be used elsewhere.

While we're at it, make the transport_wrapper be a bit more future
proof by using perfect forwarding throughoug.