Support multiple simultaneous CAN channels in python
This change largely lets you use the python library with multiple
transports simultaneously.  moteus_tool and tview will automatically
work with all connected fdcanusb and socketcan devices connected to
the system, including device discovery.  When specifying a device with
`--target`, the correct TransportDevice is detected automatically,
whether it is a number CAN ID or a UUID.