Add another layer of response filtering at the Command level
This ensures that register mode and diagnostic mode operations can
co-exist at the python API level and be sent in parallel without
interfering.

It is still possible to issue multiple queries in parallel and have
their reponses get associated with the incorrect call, however that is
probably a not very common case, and could be avoided by using
alternate source addresses.