tview: Monitor and visually highlight controllers with faults
Periodically poll each attached device to see if it reports any
faults.  If it does, flash some visual indicators so that the user
knows something is wrong.  As all position commands only check for
syntax correctness before returning OK, user's often mistakenly
believe that their commands were successful when they were not.

Stop the blinking when the fault has been "observed".  That can be
expanding "servo_stats", clicking on it if it is already expanded, or
if servo_stats.mode or servo_stats.fault are clicked.

This relies on the the moteus library properly implementing reply
filtering, so bump our version requirement there.