mpat: Implement a "minimum move time" analysis
This estimates a minimum bound on the time required to move a certain
distance starting from a stop and ending at a stop.  It assumes that
the controller can perfectly plan a trajectory that uses the peak
torque at each instant, and that acceleration and deceleration are
symmetric.  The load inertia and move distance parameterizable.

Add a tool to help estimate the inertia of a motor system.  Use it to
fill in the inertia for all supported mpat motors.