Add support for newer fdcanusb firmware
The newer firmware reports a different VID/PID and also presents as a
CDC *and* gs_usb device.  Prefer to use the CDC device if both are
available in the discovery process.

The gs_usb support in python-can doesn't support setting CAN-FD
timings, so we do not include that in the list of python-can
interfaces to auto-probe.  (The linux kernel driver does though, so
you can use it with socketcan).